+++ applications/test/PtrMap/Test-PtrMap.C
@@ -42,7 +42,7 @@ void printTable(const PtrMap<T>& table)
         Info<< iter.key() << " = ";
         if (ptr)
         {
-            Info<< *ptr << " (" << long(ptr) << ")";
+            Info<< *ptr << " (" << uintptr_t(ptr) << ")";
         }
         else
         {
